medium_wf/requirements.txt
psycopg2-binary==2.9.2
Flask
requests
functions-framework
google-cloud-error-reporting